Moved widgets and utils to separate files
