+++ lib/app/model/navigation_tab.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/cupertino.dart';
+
+class NavigationTab {
+  final Icon icon;
+  final String label;
+  final Widget child;
+  final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
+
+  NavigationTab({required this.icon, required this.label, required this.child});
+}
