+++ lib/enums/data_input_type.dart
@@ -1,0 +1,1 @@
+enum DataInputType { image, text }
