+++ lib/main.dart
@@ -1,12 +1,6 @@
-import 'dart:io';
-import 'dart:ui';
-
 import 'package:args/args.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:logging/logging.dart';
-import 'package:path/path.dart' as path_pkg;
-import 'package:path_provider/path_provider.dart';
 import 'package:provider/provider.dart';
 
 import 'app_container.dart';
@@ -14,70 +8,24 @@ import 'pages/register_page.dart';
 import 'routes.dart';
 import 'theme.dart';
 import 'ui/about_page.dart';
-import 'ui/home_page.dart';
+import 'pages/home_page.dart';
 import 'ui/new_group_page.dart';
 import 'ui/qr_reader_page.dart';
 import 'ui/search_peer_page.dart';
-
-void printUsage(ArgParser parser, IOSink sink) {
-  sink.writeln('Usage:');
-  sink.writeln(parser.usage);
-}
-
-Future<Directory> getAppDir() async {
-  if (Platform.isIOS || Platform.isMacOS) {
-    return getLibraryDirectory();
-  }
-  if (Platform.isAndroid) {
-    return getApplicationSupportDirectory();
-  }
-  final path = path_pkg.join(
-    path_pkg.dirname(Platform.resolvedExecutable),
-    'app',
-  );
-  return Directory(path);
-}
+import 'util/app_arg_parser.dart';
+import 'util/app_dir_getter.dart';
+import 'util/error_logger.dart';
 
 void main(List<String> args) async {
-  Logger.root.level = Level.WARNING;
+  // Init error logger
+  ErrorLogger().initLogger();
 
-  FlutterError.onError = (details) {
-    FlutterError.presentError(details);
-    Logger.root.severe(details.toString(), details.exception, details.stack);
-  };
-  PlatformDispatcher.instance.onError = (error, stack) {
-    Logger.root.severe(error.toString(), error, stack);
-    return false;
-  };
-
-  final parser = ArgParser()
-    ..addFlag(
-      'help',
-      abbr: 'h',
-      help: 'display usage information',
-      negatable: false,
-    )
-    ..addOption(
-      'host',
-      help: 'address of the server',
-    )
-    ..addOption(
-      'name',
-      help: 'name of the user',
-    );
-
-  ArgResults? results;
-  try {
-    results = parser.parse(args);
-    if (results['help']) printUsage(parser, stdout);
-  } on ArgParserException catch (e) {
-    stderr.writeln(e.message);
-    printUsage(parser, stderr);
-  }
+  // Parse command line arguments
+  final ArgResults argResults = AppArgParser(args: args).initParser();
 
   WidgetsFlutterBinding.ensureInitialized();
 
-  final appDir = await getAppDir();
+  final appDir = await AppDirGetter.getAppDir();
 
   runApp(
     Provider<AppContainer>(
@@ -85,19 +33,19 @@ void main(List<String> args) async {
         appDirectory: appDir,
       ),
       dispose: (_, appContainer) => appContainer.dispose(),
-      child: MyApp(
-        prefillHost: results?['host'],
-        prefillName: results?['name'],
+      child: MeeSignClient(
+        prefillHost: argResults['host'],
+        prefillName: argResults['name'],
       ),
     ),
   );
 }
 
-class MyApp extends StatelessWidget {
+class MeeSignClient extends StatelessWidget {
   final String? prefillHost;
   final String? prefillName;
 
-  const MyApp({
+  const MeeSignClient({
     super.key,
     this.prefillHost,
     this.prefillName,
@@ -105,12 +53,8 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    SystemChrome.setEnabledSystemUIMode(SystemUiMode.edgeToEdge);
-    SystemChrome.setSystemUIOverlayStyle(
-      const SystemUiOverlayStyle(
-        systemNavigationBarColor: Colors.transparent,
-      ),
-    );
+    const String defaultHost = 'meesign.crocs.fi.muni.cz';
+    configureSystemStyle();
 
     return MaterialApp(
       title: 'MeeSign',
@@ -124,12 +68,23 @@ class MyApp extends StatelessWidget {
         Routes.newGroup: (_) => const NewGroupPage(),
         Routes.newGroupSearch: (_) => const SearchPeerPage(),
         Routes.newGroupQr: (_) => const QrReaderPage(),
+        Routes.about: (_) => const AboutPage(),
         Routes.init: (_) => RegisterPage(
-              prefillHost: prefillHost ?? 'meesign.crocs.fi.muni.cz',
+              prefillHost: prefillHost ?? defaultHost,
               prefillName: prefillName ?? '',
             ),
-        Routes.about: (_) => const AboutPage(),
       },
+    );
+  }
+
+  void configureSystemStyle() {
+    // Configure system style like status bar color and navigation bar color
+    // Applies to both Android and iOS
+    SystemChrome.setEnabledSystemUIMode(SystemUiMode.edgeToEdge);
+    SystemChrome.setSystemUIOverlayStyle(
+      const SystemUiOverlayStyle(
+        systemNavigationBarColor: Colors.transparent,
+      ),
     );
   }
 }
