+++ lib/util/card_reader_launcher.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart' hide Card;
+import 'package:meesign_core/meesign_card.dart';
+import '../ui/card_reader_page.dart';
+
+void launchCardReader(
+  BuildContext context,
+  Future<void> Function(Card) onCard,
+) {
+  Navigator.push(
+    context,
+    MaterialPageRoute(
+      builder: (context) => CardReaderPage(
+        onCard: onCard,
+      ),
+    ),
+  );
+}
