+++ lib/util/list_intersperse.dart
@@ -1,0 +1,11 @@
+extension Intersperse<T> on List<T> {
+  List<T> intersperse(T inter) {
+    List<T> res = [];
+    for (T item in this) {
+      res.add(item);
+      res.add(inter);
+    }
+    if (res.isNotEmpty) res.removeLast();
+    return res;
+  }
+}
